Remove ugly if
